+++ autoapi/directives.py
@@ -24,9 +24,9 @@ class AutoapiSummary(Autosummary):  # pylint: disable=too-few-public-methods
                 if obj.overloads:
                     sig = "(\u2026)"
                 else:
-                    sig = "({})".format(obj.args)
+                    sig = f"({obj.args})"
                     if obj.return_annotation is not None:
-                        sig += " \u2192 {}".format(obj.return_annotation)
+                        sig += f" \u2192 {obj.return_annotation}"
             else:
                 sig = ""